pyproject.toml
[project]
name = "llm-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.9.0",
    "pyyaml>=6.0.3",
]
